package/uboot-tools: Add patch to fix build of uboot-tools.
The following patch adds an include of stdint.h when MTD_OLD is not
defined. The source file fw_env.c of uboot uses the types uint8_t and
uint32_t frequently. It indirectly accesses their definition through the
include of the header mtd/mtd-user.h, which in turn includes the header
stdint.h, where the above types are defined. In the newer kernel versions,
though, the header mtd/mtd-user.h does not include the header stdint.h
anymore, so the compiler does not recognize the uint8_t and uint32_t types.
The following patch to uboot-tools moves up the include of stdint.h to
affect cases where MTD_OLD is not defined to fix this issue.

Change-Id: Ie07baa17145283c1e538472f9ab305298dc3faac
Signed-off-by: Liav Rehana <liavr@mellanox.com>